Added documentation and refactored the code.
