+++ lib/pages/general/SignUpInstructionsPage.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:grade_plus_plus/pages/AbstractPage.dart';
-import 'package:grade_plus_plus/pages/fragments/StyledText.dart';
+import 'package:grade_plus_plus/pages/fragments/InstructionsStepDescription.dart';
+import 'package:grade_plus_plus/pages/fragments/DotsRow.dart';
 
+// Represents the page that contains the instructions on how to get your grades from sis
 class SignUpInstructionsPage extends AbstractPage {
   SignUpInstructionsPage({Key key})
       : super(
@@ -18,12 +20,12 @@ class _SignUpInstructionsPageState extends PageState<SignUpInstructionsPage> {
   int imgIndex = 0;
 
   final List<String> imgList = [
-    'signup_instructions1.png',
-    'signup_instructions2.png',
-    'signup_instructions3.png',
-    'signup_instructions4.png',
-    'signup_instructions5.png',
-    'signup_instructions6.png',
+    'assets/signup_instructions1.png',
+    'assets/signup_instructions2.png',
+    'assets/signup_instructions3.png',
+    'assets/signup_instructions4.png',
+    'assets/signup_instructions5.png',
+    'assets/signup_instructions6.png',
   ];
 
   final List<String> descriptionsList = [
@@ -56,13 +58,14 @@ class _SignUpInstructionsPageState extends PageState<SignUpInstructionsPage> {
             height: 10.0,
           ),
           Center(
+            // Image and dots row stack widget
             child: Stack(
               children: <Widget>[
                 Container(
                   decoration: BoxDecoration(
                     image: DecorationImage(
                       image: AssetImage(
-                        'assets/' + imgList[imgIndex],
+                        imgList[imgIndex],
                       ),
                       fit: BoxFit.contain,
                     ),
@@ -92,55 +95,18 @@ class _SignUpInstructionsPageState extends PageState<SignUpInstructionsPage> {
           SizedBox(
             height: 15.0,
           ),
+          // Buttons row
           Padding(
             padding: const EdgeInsets.only(left: 47.5, right: 47.5),
             child: Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: <Widget>[
                 if (imgIndex != 0)
-                  TextButton(
-                    onPressed: _previousImage,
-                    child: Text(
-                      'Prev',
-                      style: TextStyle(
-                        fontSize: 25.0,
-                        fontWeight: FontWeight.values[2],
-                        shadows: <Shadow>[
-                          Shadow(
-                            offset: Offset(0.0, 0.0),
-                            blurRadius: 1.5,
-                            color: Color.fromARGB(200, 0, 0, 0),
-                          ),
-                        ],
-                      ),
-                    ),
-                    style: TextButton.styleFrom(
-                      primary: Colors.blue,
-                    ),
-                  )
+                  _buildTextButton('Prev', _previousImage)
                 else
                   SizedBox(),
                 if (imgIndex != imgList.length - 1)
-                  TextButton(
-                    onPressed: _nextImage,
-                    child: Text(
-                      'Next',
-                      style: TextStyle(
-                        fontSize: 25.0,
-                        fontWeight: FontWeight.values[2],
-                        shadows: <Shadow>[
-                          Shadow(
-                            offset: Offset(0.0, 0.0),
-                            blurRadius: 1.5,
-                            color: Color.fromARGB(200, 0, 0, 0),
-                          ),
-                        ],
-                      ),
-                    ),
-                    style: TextButton.styleFrom(
-                      primary: Colors.blue,
-                    ),
-                  )
+                  _buildTextButton('Next', _nextImage)
                 else
                   SizedBox(),
               ],
@@ -153,119 +119,35 @@ class _SignUpInstructionsPageState extends PageState<SignUpInstructionsPage> {
 
   void _previousImage() {
     setState(() {
-      imgIndex = imgIndex > 0 ? imgIndex - 1 : 0;
+      imgIndex--;
     });
   }
 
   void _nextImage() {
     setState(() {
-      imgIndex = imgIndex < imgList.length - 1 ? imgIndex + 1 : imgIndex;
+      imgIndex++;
     });
   }
-}
 
-class DotsRow extends StatelessWidget {
-  final int numberOfDots;
-  final int imgIndex;
-
-  DotsRow({this.numberOfDots, this.imgIndex});
-
-  Widget _inactivePhoto() {
-    return Container(
-      child: Padding(
-        padding: const EdgeInsets.only(
-          left: 2.0,
-          right: 2.0,
-        ),
-        child: Container(
-            height: 6.0,
-            width: 6.0,
-            decoration: BoxDecoration(
-              color: Colors.grey,
-              borderRadius: BorderRadius.circular(4.0),
-            )),
-      ),
-    );
-  }
-
-  Widget _activePhoto() {
-    return Container(
-      child: Padding(
-        padding: EdgeInsets.only(
-          left: 5.0,
-          right: 5.0,
-        ),
-        child: Container(
-          height: 10.0,
-          width: 10.0,
-          decoration: BoxDecoration(
-              color: Colors.blue,
-              borderRadius: BorderRadius.circular(5.0),
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.blue,
-                  spreadRadius: 0.0,
-                  blurRadius: 2.0,
-                )
-              ]),
-        ),
-      ),
-    );
-  }
-
-  List<Widget> _buildDots() {
-    List<Widget> dots = [];
-
-    for (int i = 0; i < numberOfDots; i++)
-      dots.add(i == imgIndex ? _activePhoto() : _inactivePhoto());
-
-    return dots;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: _buildDots(),
-      ),
-    );
-  }
-}
-
-class InstructionsStepDescription extends StatelessWidget {
-  final int imgIndex;
-  final String description;
-
-  InstructionsStepDescription({this.imgIndex, this.description});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: <Widget>[
-          StyledText(
-            'Step ' + (imgIndex + 1).toString(),
-            size: 30.0,
-            weight: FontWeight.values[5],
-          ),
-          SizedBox(
-            height: 10.0,
-          ),
-          Padding(
-            padding: const EdgeInsets.only(
-              left: 35.0,
-              right: 35.0,
+  TextButton _buildTextButton(text, callback) {
+    return TextButton(
+      onPressed: callback,
+      child: Text(
+        text,
+        style: TextStyle(
+          fontSize: 25.0,
+          fontWeight: FontWeight.values[2],
+          shadows: <Shadow>[
+            Shadow(
+              offset: Offset(0.0, 0.0),
+              blurRadius: 1.5,
+              color: Color.fromARGB(200, 0, 0, 0),
             ),
-            child: StyledText(
-              description,
-              align: TextAlign.center,
-              size: 18.0,
-              weight: FontWeight.values[4],
-            ),
-          )
-        ],
+          ],
+        ),
+      ),
+      style: TextButton.styleFrom(
+        primary: Colors.blue,
       ),
     );
   }
